+++ jena/src/main/java/org/apache/commons/rdf/jena/impl/JenaVariableImpl.java
@@ -24,11 +24,11 @@ import org.apache.commons.rdf.jena.JenaRDFTerm;
 import org.apache.commons.rdf.jena.JenaVariable;
 import org.apache.jena.graph.Node;
 
-public class VariableImpl implements JenaRDFTerm, JenaVariable {
+public class JenaVariableImpl implements JenaRDFTerm, JenaVariable {
 
 	private Node node;
 
-	VariableImpl(Node node) {	
+	JenaVariableImpl(Node node) {	
 		if (! node.isVariable()) {
 			throw new IllegalArgumentException("Node is not a variable: " + node);
 		}
